feat: add styled-components.📚 refactor: change directory structure.🔧
