+++ package.json
@@ -35,5 +35,9 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/styled-components": "^4.4.2",
+    "styled-components": "^5.0.0"
   }
 }
